Avoid auto complete for non-local databases

+++ src/main/java/jacz/store/database/models/Chapter.java
@@ -1,5 +1,6 @@
 package jacz.store.database.models;
 
+import jacz.store.database.DatabaseMediator;
 import org.javalite.activejdbc.Model;
 import org.javalite.activejdbc.annotations.BelongsTo;
 
@@ -13,17 +14,21 @@ public class Chapter extends Model {
 
     @Override
     public void beforeDelete() {
-        // delete people association records
-        ChaptersPeople.deleteRecords("chapter_id", getId());
-        // delete video files
-        ChaptersVideoFiles.deleteRecords("chapter_id", getId());
-        DeletedItem.addDeletedItem(this, getTableName());
+        if (DatabaseMediator.mustAutoComplete()) {
+            // delete people association records
+            ChaptersPeople.deleteRecords("chapter_id", getId());
+            // delete video files
+            ChaptersVideoFiles.deleteRecords("chapter_id", getId());
+            DeletedItem.addDeletedItem(this, getTableName());
+        }
     }
 
     static void deleteRecords(Model baseModel) {
-        List<Chapter> chapterModels = baseModel.getAll(Chapter.class);
-        for (Chapter chapter : chapterModels) {
-            chapter.delete();
+        if (DatabaseMediator.mustAutoComplete()) {
+            List<Chapter> chapterModels = baseModel.getAll(Chapter.class);
+            for (Chapter chapter : chapterModels) {
+                chapter.delete();
+            }
         }
     }
 }
